+++ pkg/kubelet/cm/BUILD
@@ -31,6 +31,7 @@ go_library(
     }),
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/api/v1/helper:go_default_library",
         "//pkg/features:go_default_library",
         "//pkg/kubelet/apis/cri:go_default_library",
         "//pkg/kubelet/apis/deviceplugin/v1alpha1:go_default_library",
@@ -52,7 +53,6 @@ go_library(
     ] + select({
         "@io_bazel_rules_go//go/platform:linux_amd64": [
             "//pkg/api:go_default_library",
-            "//pkg/api/v1/helper:go_default_library",
             "//pkg/api/v1/helper/qos:go_default_library",
             "//pkg/api/v1/resource:go_default_library",
             "//pkg/kubelet/cm/util:go_default_library",
