+++ pkg/kubelet/deviceplugin/endpoint_test.go
@@ -114,6 +114,13 @@ func TestListAndWatch(t *testing.T) {
 
 }
 
+func TestCloneDevice(t *testing.T) {
+	d := cloneDevice(&pluginapi.Device{ID: "ADeviceId", Health: pluginapi.Healthy})
+
+	require.Equal(t, d.ID, "ADeviceId")
+	require.Equal(t, d.Health, pluginapi.Healthy)
+}
+
 func esetup(t *testing.T, devs []*pluginapi.Device, socket, resourceName string, callback MonitorCallback) (*Stub, *endpoint) {
 	p := NewDevicePluginStub(devs, socket)
